refactor(keystore): apply cleanup review findings
- Share `createJsonStorage` from test/utils.ts instead of three per-file
  copies (Store.test.ts already had one)
- Replace the `prepared.key!` assertion with an explicit guard
- Build the access-key record from a single material selection in `add`
- Make the hydrate cache lookup an explicit get/set instead of `?? (iife)()`
