+++ test/utils.ts
@@ -1,9 +1,30 @@
 import * as Http from 'node:http'
 import type { AddressInfo } from 'node:net'
+import { Json } from 'ox'
+
+import * as Storage from '../src/core/Storage.js'
 
 export type Server = Http.Server & {
   closeAsync: () => Promise<unknown>
   url: string
+}
+
+/** String-based storage adapter (values survive only as JSON strings). */
+export function createJsonStorage() {
+  const items = new Map<string, string>()
+  return Storage.from({
+    getItem<value>(name: string) {
+      const value = items.get(name)
+      if (!value) return null
+      return Json.parse(value) as value
+    },
+    removeItem(name: string) {
+      items.delete(name)
+    },
+    setItem(name: string, value: unknown) {
+      items.set(name, Json.stringify(value))
+    },
+  })
 }
 
 export function createServer(
